requirements.txt
flask>=3.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
line-bot-sdk>=3.5.0
openai>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.2
certifi>=2023.0.0

